Switch to ncm2 (+ language servers) for completion
For Python, that means replacing jedi-vim. This makes it much easier to
get completion for additional languages (basically just install their
language server and configure its launch command in .vimrc).

ncm2 completion targets Neovim. It can be made to work on Vim 8, but
here's a rule of thumb: if you can't be bothered to set up Neovim on a
machine, you probably don't need completion there.