Update nvim-treesitter config
Indentation support has landed! The definitions for Python don't exist
yet, but I'm still getting rid of the separate Python indentation plugin
to remind myself that I should check. Also, since I'm using Black for
auto-formatting, the ugly default indentation is less of an issue now.

The Treesitter playground is a really nice way to visualize the parsed
tree for a given buffer. It will teach me about parsing and help if I
ever need to create my own Treesitter queries.

Basic Markdown support is finally available, but I should probably still
wait for languagetree to settle down. Also, the Markdown parser
currently seems to be responsible for a crash:

- https://github.com/nvim-treesitter/nvim-treesitter/issues/602.
- https://github.com/ikatyang/tree-sitter-markdown/issues/14

So not getting rid of the Markdown syntax plugin just yet, and disabling
the markdown parser for now.

And last but not least, using the new ensure_installed option to install
the parsers.