Change LSP go-to-definition mapping
`C-]` is hard to use with a Czech keyboard layout in Windows Terminal, plus in
general I don't really like chords, even though this is a standard one. `gd`
also happens to be a standard mapping meaning "go to definition" (although not
using tags), so remapping to something conceptually similar but better makes
sense. And it's much nicer to type.