Remove iron.nvim #lua
Updates broke my config, which brought it to my attention that even if I
fix my config, I can't get an iron REPL to start. I don't really have
time to debug this, plus I didn't use iron REPLs that much anyway -- my
typical workflow is to open IPython in a separate terminal. So I'm just
getting rid of all this, no hard feelings.

In the future (maybe when 0.5 stabilizes and ships?), it might be worth
rechecking a plugin with this type of functionality. Maybe not
necessarily iron, Conjure looks like it might get Python support once
Treesitter lands in stable.

At any rate, this commit removes a lot of Lua code which might be
helpful for future reference, so I'm marking this commit with #lua as
containing interesting material for customizing Neovim with Lua in the
future. Hopefully I'll learn to use these hashtags consistently, it
seems like a nice way to point out commits which remove potentially
useful material.