Revamp path setup in config.fish
For one thing, the update in e4c501f to account for the new pyenv behavior
w.r.t. (not) modifying the path wasn't exactly correct. For another, Homebrew
now behaves similarly. Plus I needed to change the Homebrew config anyway
because of switching to an M1 Mac.

Previously, in order to play nice with subshells, I'd check the path when
adding a new element to see whether it already contained it, so as not to add
duplicates. This new approach is simpler -- it just checks a sentinel variable.
If it's set, then the corresponding path modification has already been
performed in a supershell and it's skipped.