Searching for a fuzzy finder #tmp
Probably Telescope. It looks nice while integrating with my theme, the
UI is mostly legible without requiring tweaks, it also has LSP-related
functionality by default, and it's all Lua so it's easy to install and
harder to break (all it needs to run is already right there in Neovim).
There are reportedly some performance issues when searching a lot of
files, but it looks like it's one of the key issues they're trying to
address right now.