Bump default max_line_length to 88
It offers a bit more wiggle room than 72 with long variable names, URLs
etc. And Python files get auto-formatted by Black anyway, so 72 was just
an annoying limitation that would be to eager to start a new line that
I'd have to erase and Black would format over it anyway.