Org-roam: Tweak capture templates
I'm using the default daily template, so don't copy it to my config so
that I get upstream changes automatically.

Instead, I made a "tagged" capture template which tried to be moderately
fancy and took me a long time to get right (also because Org-roam does
some preprocessing of its own before handing off to vanilla Org
capture), so that might be useful for future reference.

Essentially: don't try to add newlines in the template itself, it's
about using the correct template type (e.g. item vs. plain) and the
:empty-lines* options.