Emacs: PGTK dead keys and Doom leader alt key
Turns out, for dead keys to work with PGTK, you need input methods
enabled, which sort of makes sense, PGTK brings Emacs in line with
desktop standards and input methods is the standard way to solve this.

(Without PGTK, the solution was actually the reverse -- to *disable*
input methods by unsetting XMODIFIERS.)

But that means you can't use S-SPC as doom-leader-alt-key, because for
that to work, input methods need to be disabled via setting
pgtk-use-im-context-on-new-connection to nil.

(I wonder how on earth I thought I don't need "to type accented
characters that aren't directly on [my] keyboard". In retrospect, it's
pretty obvious I do.)

So I'll be leaving input methods enabled with PGTK, and sticking to the
standard M-SPC as doom-leader-alt-key, which just means I need to
disable this shortcut in GNOME.