Doom: Don't override blocks with Jupyter
The primary reason was that it's shorter to read and type python or R
than jupyter-python or jupyter-R. But for the typing, you've got tempo
templates. For the reading, you can live with it.

But most importantly, the override was sort of brittle, you had to
figure out a very workaroundy way to do it give Doom's lazy loading, and
based on Doom's Jupyter-related modules, Doom relies on the blocks' lang
being prefixed with jupyter- to implement its completion and
documentation lookup functionality for these blocks.

So better stick with the defaults if it means better support
out-of-the-box instead of maintaining an elaborate tower of hacks and
workarounds just because I think it's cooler not to have that jupyter-
prefix.

Also, if I ever need to test or play with vanilla Python Org blocks,
this makes it a no-brainer.