Emacs Jupyter: Don't prettify :text/plain
I thought using :text/plain was enough in itself to prevent
prettification into Org tables. Turns out it isn't, for reasons that are
reasonable in retrospect -- prettification is client-side in vanilla
Org, so Emacs Jupyter tries to do that by default as well. Only if a
kernel returns a pre-prettified (on the server side) :text/org mime type
does it use that directly without any further client-side processing.