Zotero: More resilient BBT postscript
If a.itemKey is undefined, default to a.key. The null coalescing
operator ?? wouold be more correct than ||, but BBT highlights the
postscript box in orange, signifying an error, and while it apparently
works on Linux, it doesn't on macOS. And since I don't need to
distinguish between a.itemKey being undefined vs. generally falsey, ||
is probably fine.